requirements.txt
pandas==2.2.2
numpy==1.26.4
torch==2.5.1
torchaudio==2.5.1
transformers==4.43.3
pyannote.audio==3.3.2
boto3==1.39.4
python-dotenv==1.0.1
joblib==1.4.2
tenacity==9.1.2
psutil==5.9.0
huggingface-hub==0.30.2
rapidfuzz==3.12.2
psycopg2-binary==2.9.10
uuid==1.30
openpyxl==3.1.5
minio==7.2.16


